pyproject.toml
[tool.poetry]
name = "my-langgraph-project"
version = "0.1.0"
description = ""
authors = ["최준우 <emotree@wavve.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "streamlit run main.py"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# 언어 및 자연어 처리 관련 패키지
langchain = "^0.3"
langchain-core = "^0.3"
langchain-experimental = "^0.3"
langchain-community = "^0.3"
langchain-openai = "^0.2"
langchain-anthropic = "^0.2"
langchain-text-splitters = "^0.3"
langchain-elasticsearch = "^0.3"
langchain-upstage = "^0.3"
langchain-chroma = "^0.1"
langchain-cohere = "^0.3"
langchain-milvus = "^0.1.1"
langchain-google-genai = "^2.0"
langchain-huggingface = "^0"
langchain-teddynote = "^0.3.37"
langchainhub = "^0.1.15"
langgraph = "^0"
langsmith = "^0.1"
openai = "^1.34"
deepl = "^1.17.0"
kiwipiepy = "^0.17.1"
konlpy = "^0.6.0"

# 데이터 처리 및 분석 관련 패키지
pandas = "^2.2.2"
rank-bm25 = "^0.2.2"

# 시각화 및 사용자 인터페이스 관련 패키지
matplotlib = "^3.8.4"
streamlit = "^1.32.2"
jupyter = "^1.0.0"
notebook = "^7.1.2"

# 기타 유틸리티 및 필수 패키지
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
lxml = "^5.2.2"
pillow = "^10.3.0"

google-api-python-client = "^2.162.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
